# docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  controllers.customerDTO:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        maxLength: 15
        type: string
      last_name:
        maxLength: 15
        type: string
      password:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      username:
        maxLength: 16
        type: string
    required:
    - address
    - city
    - country
    - email
    - first_name
    - last_name
    - phone
    - postal_code
    - username
    type: object
...
  controllers.updateCustomerDTO:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        maxLength: 15
        type: string
      last_name:
        maxLength: 15
        type: string
      phone:
        type: string
      postal_code:
        type: string
    required:
    - address
    - city
    - country
    - email
    - first_name
    - last_name
    - phone
    - postal_code
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /customer/login:
    post:
      description: Logging in to get jwt token to access customer api by roles.
      parameters:
      - description: the body to login a customer
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as customer.
      tags:
      - Auth
...
  /customers/{id}:
    delete:
      description: Delete a Customer by id.
      parameters:
      - description: Customer id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete one Customer.
      tags:
      - Customer
    get:
      description: Get a Customer by id.
      parameters:
      - description: Customer id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Customer.
      tags:
      - Customer
    put:
      description: Update Customer by id.
      parameters:
      - description: Customer id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an Customer
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.updateCustomerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Customer.
      tags:
      - Customer
  /customers/{id}/products:
    get:
      description: Get all Products by Customer Id.
      parameters:
      - description: Customer id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Products.
      tags:
      - Customer
  /customers/register:
    post:
      description: Creating a new Customer.
      parameters:
      - description: the body to create a new Customer
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.customerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register New Customer.
      tags:
      - Customer
...
  /seller/login:
    post:
      description: Logging in to get jwt token to access seller api by roles.
      parameters:
      - description: the body to login a seller
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login as seller.
      tags:
      - Auth
...
swagger: "2.0"